[Sounds] Add microphone functions and typedocs.
@types/dom-mediacapture-record added to tsconfig.json
Extraneous Test bundle and tab deleted.
