--- .prettierignore
@@ -1,4 +1,0 @@
-# Files to ignore #
-###################
-/src/bundles/sounds/index.ts
-/src/tabs/Sounds/index.tsx
