+++ src/bundles/sounds/types.ts
@@ -6,7 +6,7 @@ export type List = EmptyList | NonEmptyList;
 export type Wave = (...t: any) => number;
 export type Sound = Pair<Wave, number>;
 export type SoundProducer = (...t: any) => Sound;
-export type UnaryComposition = (s: Sound) => Sound;
+export type SoundTransformer = (s: Sound) => Sound;
 export type AudioElement = HTMLAudioElement;
 export type ErrorLogger = (
   error: string | string[],
