--- src/bundles/test/index.ts
@@ -1,12 +1,0 @@
-// import __Params from '../../typings/__Params';
-
-function test() {
-  return 'test';
-}
-
-// exported functions
-export default function () {
-  return {
-    test,
-  };
-}
